(APWA-21) [codegen.yml, Product, asyncActions, useCustomProduct, useProductImageCarousel, useSignIn, useSignOut, useLoadUser, useUserSession, useWishlist, useWishlistItem]
- Refactored talons (Wishlist, User, SignOut, SignIn, ProductImageCarousel).
