+++ src/peregrine/lib/talons/adeoweb/ProductImageCarousel/useProductImageCarousel.ts
@@ -3,13 +3,31 @@ import { useCallback, useEffect } from 'react';
 import { useCarousel } from '@magento/peregrine';
 
 import { PREFETCH_IMAGES } from 'src/lib/constants/swMessageTypes';
+import { TMediaGalleryInterface } from 'src/lib/types/graphql/MediaGalleryInterface';
 import { generateUrlFromContainerWidth } from 'src/lib/util/images';
 import {
     sendMessageToSW,
     VALID_SERVICE_WORKER_ENVIRONMENT
 } from 'src/lib/util/swUtils';
 
-export const useProductImageCarousel = props => {
+type TUseProductImageCarousel = {
+    currentImage: TMediaGalleryInterface;
+    activeItemIndex: number;
+    altText: string;
+    handleNext: () => void;
+    handlePrevious: () => void;
+    handleThumbnailClick: (index: number) => void;
+    sortedImages: TMediaGalleryInterface[];
+};
+
+interface IUseProductImageCarouselProps {
+    images: TMediaGalleryInterface[];
+    imageWidth: number;
+}
+
+export const useProductImageCarousel = (
+    props: IUseProductImageCarouselProps
+): TUseProductImageCarousel => {
     const { images = [], imageWidth } = props;
     const [carouselState, carouselApi] = useCarousel(images);
     const { activeItemIndex, sortedImages } = carouselState;
@@ -30,6 +48,8 @@ export const useProductImageCarousel = props => {
     useEffect(() => {
         if (VALID_SERVICE_WORKER_ENVIRONMENT) {
             const urls = images.map(
+                // TODO: `file` does not exist on TMediaGalleryInterface. Refactor this hook later.
+                // @ts-expect-error
                 ({ file }) =>
                     `${location.origin}${generateUrlFromContainerWidth(
                         file,
