--- src/peregrine/lib/talons/adeoweb/SignIn/useSignIn.d.ts
@@ -1,19 +1,0 @@
-import { DocumentNode } from 'graphql';
-
-import { TGenerateCustomerTokenProps } from 'src/lib/types/graphql/SignIn';
-
-type TUseSignIn = {
-    handleSubmit: (props: TGenerateCustomerTokenProps) => void;
-    isBusy: boolean;
-    isSignedIn: boolean;
-};
-
-export type TUseSignInProps = {
-    signInMutation: DocumentNode;
-    customerQuery: DocumentNode;
-    getCustomerCartQuery: DocumentNode;
-    mergeCartsMutation: DocumentNode;
-    signOutMutation: DocumentNode;
-};
-
-export function useSignIn(props: TUseSignInProps): TUseSignIn;
