--- src/peregrine/lib/talons/adeoweb/SignOut/useSignOut.d.ts
@@ -1,12 +1,0 @@
-import { DocumentNode } from 'graphql';
-
-type TUseSignOutProps = {
-    signOutMutation: DocumentNode;
-    createCartMutation: DocumentNode;
-};
-
-export type TUseSignOut = {
-    handleSignOut: () => void;
-};
-
-export function useSignOut(props: TUseSignOutProps): TUseSignOut;
