--- src/peregrine/lib/talons/adeoweb/User/useCurrentUserData.d.ts
@@ -1,7 +1,0 @@
-import { TCustomer } from 'src/lib/types/graphql/Customer';
-
-type TUseLoadUser = {
-    currentUser: TCustomer;
-};
-
-export function useCurrentUserData(): TUseLoadUser;
