--- src/peregrine/lib/talons/adeoweb/User/useCurrentUserData.js
@@ -1,9 +1,0 @@
-import { useUserContext } from 'src/peregrine/lib/context/adeoweb/user';
-
-export const useCurrentUserData = () => {
-    const [{ currentUser }] = useUserContext();
-
-    return {
-        currentUser
-    };
-};
