--- src/peregrine/lib/talons/adeoweb/User/useLoadUser.d.ts
@@ -1,11 +1,0 @@
-import { DocumentNode } from 'graphql';
-
-type TUseLoadUserProps = {
-    getUserDetailsQuery: DocumentNode;
-};
-
-type TUseLoadUser = {
-    error: null | Error;
-};
-
-export function useLoadUser(props: TUseLoadUserProps): TUseLoadUser;
