--- src/peregrine/lib/talons/adeoweb/User/useUserSession.d.ts
@@ -1,7 +1,0 @@
-import { DocumentNode } from 'graphql';
-
-type TUseUserSessionProps = {
-    signOutMutation: DocumentNode;
-};
-
-export function useUserSession(props: TUseUserSessionProps): void;
