+++ src/peregrine/lib/talons/adeoweb/User/useUserSession.ts
@@ -1,12 +1,18 @@
+import { DocumentNode } from 'graphql';
+
 import { useMutation } from '@apollo/react-hooks';
 import { useEffect } from 'react';
 
 import { useUserContext } from 'src/peregrine/lib/context/adeoweb/user';
 import { fetchPolicy } from 'src/peregrine/lib/util/adeoweb/fetchPolicy';
 
-export const useUserSession = ({ signOutMutation }) => {
-    const userSes = useUserContext();
+type TUseUserSessionProps = {
+    signOutMutation: DocumentNode;
+};
 
+export const useUserSession = ({
+    signOutMutation
+}: TUseUserSessionProps): void => {
     const [, { setUserSession }] = useUserContext();
 
     const [revokeToken] = useMutation(signOutMutation, {
